Migrated ci from Travis CI to GitHub Actions workflow.
This repository is currently using https://travis-ci.org/. However, https://travis-ci.org/ was terminated on June 15, 2021.

> Since June 15th, 2021, the building on travis-ci.org is ceased. Please use travis-ci.com from now on.
> https://travis-ci.org/github/railsware/js-routes

Therefore, I moved to GitHub Action to restart CI.

The names of environment variables for CI determination changed with the transition to github action.

> Default environment variables
>
> | Environment variable | Description         |
> | -------------------- | ------------------- |
> | CI                   | Always set to true. |
>
> https://docs.github.com/en/actions/learn-github-actions/environment-variables#default-environment-variables

For the version of Ruby and related libraries that run CI, we try to use non-EOL versions.

Ruby

> Ruby 2.7
> status: security maintenance
> release date: 2019-12-25
> https://www.ruby-lang.org/en/downloads/branches/

Rails

> 4 Severe Security Issues
> Currently included series: 7.0.Z, 6.1.Z, 6.0.Z, 5.2.Z.
> https://guides.rubyonrails.org/maintenance_policy.html#severe-security-issues

Previously, jruby and head were specified with allow_failures, but GitHub Action does not support allow_failures, so they were removed from the target.