Obfuscate assignment to module.exports to prevent warnings
Some javascript bundles (like vite) emit warnings when they encounter
module.export in an ESM module. This just hides that assignment a little
so that that warning doesn't appear.

The spec is a little odd, as I wanted to avoid rewriting the comment in
the routes file to avoid saying module.exports, so it strips out
single-line javascript comments.